pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oncall-env"
version = "0.1.0"
description = "OnCallEnv OpenEnv environment for SRE incident response"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "openai>=1.0.0",
    "pydantic>=2.6.0",
    "PyYAML>=6.0.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[tool.setuptools]
include-package-data = true
packages = ["env", "server"]
